+++ NEWS.md
@@ -7,6 +7,9 @@ Install the development version with: `install_github("rstudio/reticulate")`
 
 - Add `Library\bin` to PATH on Windows to ensure Anaconda can find MKL
 
+- New `source_python()` function for sourcing Python scripts into R 
+  environments.
+
 
 ## reticulate 1.4 (CRAN)